camelCase, GetDay opt, db opt

+++ src/db.ts
@@ -12,16 +12,33 @@ const handleError = (error: mongoose.Error) => console.log("몽고디비 연결 
 db.once("open", handleOpen);
 db.on("error", handleError);
 
-async function ClearCollections() {
-  const scheDocs = mongoose.connection.collections.schedules;
-  if (scheDocs) {
-    scheDocs.drop().then(() => console.log("schedules collection dropped"));
+async function clearCollection(colname: string) {
+  if (colname === "schedules") {
+    const scheDocs = mongoose.connection.collections.schedules;
+    if (scheDocs) {
+      scheDocs.drop().then(() => console.log("schedules collection dropped"));
+    }
   }
-
-  const idDocs = mongoose.connection.collections.id2nicks;
-  if (idDocs) {
-    idDocs.drop().then(() => console.log("id2nicks collection dropped"));
+  if (colname === "id2nicks") {
+    const idDocs = mongoose.connection.collections.id2nicks;
+    if (idDocs) {
+      idDocs.drop().then(() => console.log("id2nicks collection dropped"));
+    }
   }
 }
 
-export default ClearCollections;
+async function isColExists(colname: string) {
+  if (colname === "schedules") {
+    const scheDocs = mongoose.connection.collections.schedules;
+    if (scheDocs) return true;
+    return false;
+  }
+  if (colname === "id2nicks") {
+    const idDocs = mongoose.connection.collections.id2nicks;
+    if (idDocs) return true;
+    return false;
+  }
+  return false;
+}
+
+export { clearCollection, isColExists };
